+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added support for query `PeerList`
 - Added support for query `PeerStats`
 - Added support for query `PeerVersion`
+- Added support for query `GetAncestors`
 - Added support for query `NodeInfo`
 - Added support for query `JoinNetwork`
 - Added support for query `LeaveNetwork`